+++ app/(payload)/admin/importMap.ts
@@ -1,8 +1,10 @@
+import { AdminDashboard as AdminDashboard_ccba4cbdf645413d612d3ec9b423815f } from '@/app/(payload)/components/AdminDashboard'
 import { VercelBlobClientUploadHandler as VercelBlobClientUploadHandler_16c82c5e25f430251a3e3ba57219ff4e } from '@payloadcms/storage-vercel-blob/client'
 import { CollectionCards as CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1 } from '@payloadcms/next/rsc'
 
 /** @type import('payload').ImportMap */
 export const importMap = {
+  "@/app/(payload)/components/AdminDashboard#AdminDashboard": AdminDashboard_ccba4cbdf645413d612d3ec9b423815f,
   "@payloadcms/storage-vercel-blob/client#VercelBlobClientUploadHandler": VercelBlobClientUploadHandler_16c82c5e25f430251a3e3ba57219ff4e,
   "@payloadcms/next/rsc#CollectionCards": CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1
 }